requirements.txt
biopython==1.74
bresenham==0.2.1
Click==7.0
cycler==0.10.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.17.3
pandas==0.25.3
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
scipy==1.3.1
six==1.12.0
Werkzeug==0.16.0
yapf==0.28.0
